optimized bundle size by updating split points
