+++ frontend/src/components/MarkdownView.js
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { helpers } from '../common/helpers';
+import { markdownConverter } from '../common/markdown';
 
 export const MarkdownView = ({ className, markdown, ...props }) => (
   <div
     className={classNames('oh-markdown-view', className)}
     {...props}
-    dangerouslySetInnerHTML={{ __html: helpers.markdownConverter.makeHtml(markdown) }}
+    dangerouslySetInnerHTML={{ __html: markdownConverter.makeHtml(markdown) }}
   />
 );
