+++ frontend/src/services/operatorsService.js
@@ -1,8 +1,6 @@
 import axios from 'axios';
-import * as _ from 'lodash-es';
 import { helpers } from '../common/helpers';
 import { reduxConstants } from '../redux';
-import { mockOperators } from '../__mock__/operators';
 
 const serverHost = process.env.DEV_HOST || 'localhost';
 const serverPort = process.env.DEV_PORT || 8080;
@@ -16,15 +14,6 @@ export const fetchOperator = (operatorName, packageName, channel) => dispatch =>
   dispatch({
     type: helpers.PENDING_ACTION(reduxConstants.GET_OPERATORS)
   });
-
-  if (process.env.MOCK_MODE) {
-    const operator = _.find(mockOperators, { name: operatorName });
-    dispatch({
-      type: helpers.FULFILLED_ACTION(reduxConstants.GET_OPERATOR),
-      payload: operator
-    });
-    return;
-  }
 
   const config = {
     params: {
@@ -53,14 +42,6 @@ export const fetchOperators = () => dispatch => {
   dispatch({
     type: helpers.PENDING_ACTION(reduxConstants.GET_OPERATORS)
   });
-
-  if (process.env.MOCK_MODE) {
-    dispatch({
-      type: helpers.FULFILLED_ACTION(reduxConstants.GET_OPERATORS),
-      payload: _.cloneDeep(mockOperators)
-    });
-    return;
-  }
 
   axios
     .get(allOperatorsRequest)
